marked-for-deletion/v00-multi-agent-expertise-framework/requirements.txt
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0
pyyaml>=6.0
jinja2>=3.1.2
python-dotenv>=1.0.0
httpx>=0.24.0
jsonschema>=4.17.0
